Move standard editor to its own directory
