+++ web/packages/teleport/src/Roles/RoleEditor/StandardEditor/StandardEditor.tsx
@@ -46,13 +46,12 @@ import {
 import { SlideTabs } from 'design/SlideTabs';
 import { RadioGroup } from 'design/RadioGroup';
 import Select from 'shared/components/Select';
-
 import { components, MultiValueProps } from 'react-select';
-
 import { Role, RoleWithYaml } from 'teleport/services/resources';
 import { LabelsInput } from 'teleport/components/LabelsInput';
+import { FieldMultiInput } from 'shared/components/FieldMultiInput/FieldMultiInput';
 
-import { FieldMultiInput } from '../../../../shared/components/FieldMultiInput/FieldMultiInput';
+import { EditorSaveCancelButton } from '../Shared';
 
 import {
   roleEditorModelToRole,
@@ -96,7 +95,6 @@ import {
   WindowsDesktopSpecValidationResult,
   AccessRuleValidationResult,
 } from './validation';
-import { EditorSaveCancelButton } from './Shared';
 import { RequiresResetToStandard } from './RequiresResetToStandard';
 
 export type StandardEditorProps = {
