+++ web/packages/teleport/src/Roles/RoleEditor/StandardEditor/withDefaults.ts
